Ceil numbers parsing & Token status
Добавил токенам статусы (валидные и с ошибками), для токенов которые фактически являются верными, но нарушают внутреннее условие, которое не влияет на внешнюю запись токена (например запись числа, превышающего максимальное значение его типа).
Добавил парсер всех целочисленных типов (Reader + набор параметров к нему)
Также заменил интерфейс ITextSource на абстрактный класс TextSource, чтобы избежать проблем при структур (боксинг и копирование элемента, что может привезти к бесконечным циклам)

+++ Text/Readers/ObjectReader/ObjectReader_Integers.cs
@@ -23,7 +23,7 @@ namespace Zion
             (A, B) => (sbyte)(A / B),
             (A, B) => (sbyte)(A % B),
             (A, B) => (sbyte)(A << B),
-            (A, B) => (sbyte)((int)A | B)
+            (A, B) => (sbyte)(A | B)
         );
 
         private static readonly IntReaderData<BigInt> BigIntData = new
@@ -45,7 +45,7 @@ namespace Zion
             (A, B) => (short)(A / B),
             (A, B) => (short)(A % B),
             (A, B) => (short)(A << B),
-            (A, B) => (short)((int)A | B)
+            (A, B) => (short)(A | B)
         );
         private static readonly IntReaderData<int> Int32Data = new
         (
